fix: update memory management by refining BSS initialization and string conversion
