+++ kernel/src/main.rs
@@ -158,7 +158,7 @@ pub mod device;
 pub mod fs;
 
 extern crate alloc;
-use alloc::string::String;
+use alloc::string::ToString;
 use device::{fdt::{init_fdt, relocate_fdt}, manager::DeviceManager};
 use initcall::{driver::driver_initcall_call, early::early_initcall_call, initcall_task};
 
@@ -168,24 +168,8 @@ use arch::init_arch;
 use task::new_kernel_task;
 use vm::kernel_vm_init;
 use sched::scheduler::get_scheduler;
-use mem::allocator::init_heap;
+use mem::{allocator::init_heap, init_bss};
 use timer::get_kernel_timer;
-
-
-pub fn init_bss() {
-    unsafe extern "C" {
-        static mut __BSS_START: u8;
-        static mut __BSS_END: u8;
-    }
-
-    unsafe {
-        let bss_start = &raw mut __BSS_START as *mut u8;
-        let bss_end = &raw mut __BSS_END as *mut u8;
-        let bss_size = bss_end as usize - bss_start as usize;
-        core::ptr::write_bytes(bss_start, 0, bss_size);
-    }
-}
-
 
 #[cfg(test)]
 pub mod test;
@@ -240,7 +224,7 @@ pub extern "C" fn start_kernel(cpu_id: usize) -> ! {
     let scheduler = get_scheduler();
     /* Make idle task as initial task */
     println!("[Scarlet Kernel] Creating initial kernel task...");
-    let mut task = new_kernel_task(String::from("Initcall"), 0, initcall_task);
+    let mut task = new_kernel_task("Initcall".to_string(), 0, initcall_task);
     task.init();
     scheduler.add_task(task, cpu_id);
     println!("[Scarlet Kernel] Scheduler will start...");
